refactor (nextjs): extract dynamic zone component logic
